nicotine-plus: 1.4.1 -> unstable-2020-05-24
Includes port to Python 3 and GTK+ 3/PyGObject; see
https://github.com/Nicotine-Plus/nicotine-plus/pull/106 for details.

This unstable version identifies itself as 1.4.3, and is planned to be
released as 2.0.0 after it has had more fixes and testing through the
official upstream PPA. The codebase is old and flaky to begin with,
and there have been several unrelated bug fixes in the years since
the last stable release, so I think it's worth updating to be able
to use non-obsolescent versions of the core dependencies, and have
nixos-unstable users help shake out any remaining issues with the port.
I verified that searching and downloading seem to work fine.

Drop the rename of the executable to `nicotine-plus`, as the test
depends on the executable name and there doesn't seem to be any strong
reason to diverge from upstream here (the original Nicotine is long
dead). This is technically a breaking change, so it might be best to
include a symlink at `$out/bin/nicotine-plus` temporarily.