acme-client: 0.2.5 -> 1.0.0
The upstream version fails to compile due to a missing limits.h include.
I added a patch to fix that.

I opened a pull request to upstream it too, but the project has moved
from GitHub onto sr.ht and now asks me to send a patch to the mailing
list. My default email client is not really suitable for that, and
getting git-send-email set up will take some work, so in the meantime
it is easier to just patch it here.