gnucash: Add libdbiDrivers env var to wrapper
The GNU Cash package clearly *wants* to be able to use the libdbi
backends, since it pulls them in as dependencies.  However, you can only
open xml formatted GNU cash files.

The CMake scripts hard-code the DVD path to be basically
<PATH_OF_LIBDBI>/dbd.  However GNU Cash does check the environment
variable GNC_DBD_DIR, so I set that environment variable in the wrapper
script.

With this change, you should be able to e.g. "Save As" in the sqlite
format.

Fixes issue #57445