podman-compose: Don't wrap podman
Podman is now a wrapped with it's container runtimes in the podman
closure.

This means that the podman passed to podman-compose is not necessarily
the one that the user wants to use as it lacks user configured
container runtimes.

This change aims to make package composition more intuitive by
defering the installation of podman to the user of podman-compose.