wireguard-go: fix executable name
It's supposed to be `wireguard-go` instead of `wireguard`. Upstream does
this right in their Makefile, however we use our own build-script which
creates a wrong file in $out, so it has to be fixed in the
`postInstall`-hook.

Closes #88567

(cherry picked from commit 0f65693e6b63b47c89f5a77cf6b985eb652485a7)