acme-client: 1.0.0 -> 1.0.1
I submitted my patch upstream, it was merged, and version 1.0.1 which
includes it has been released. So the patch here is no longer necessary.

Also, the maintainer added a Nixpkgs-based build environment to the
upstream CI setup, so in the future the upstream version will likely
not need any patching.