Calculate sha256 on the fly on dockerTools.buildLayeredImage
Previously, we were calculating layer checksums at a separate pass after
the creation of the tarball. This is not ideal when packaging containers
with large layers, since it requires reading the whole tarball again
from disk in order to create the checksum.

This commit calculates the sha256 in parallel when creating the tarball
and saves it to a file for later use by buildLayeredImage function.