sbcl: add option to disable immobile space
Programs which generate and compile a lot of code at runtime (such as
programming language interpreters like ACL2) are not suited for running on SBCL
executables built with the "immobile space" feature, as explained by Douglas
Katzman in this mail thread:

  https://sourceforge.net/p/sbcl/mailman/message/36007057/

In this commit, I add an optional flag to the SBCL package allowing you to
disable the "immobile space" features.

I also migrated away from specifying enabled/disabled features in a
`customize-target-features.lisp` file and towards supplying them as command line
arguments to `make.sh`, as has been recommended by the installation instructions
since 2012 or so.