.github/stale.yml: fix formatting
Before this, the message being written is:
```
Thank you for your contributions.
This has been automatically marked as stale because it has had no activity for 180 days.
If this is still important to you, we ask that you leave a comment below. Your comment can be as simple as "still important to me". This lets people see that at least one person still cares about this. Someone will have to do this at most twice a year if there is no other activity.
Here are suggestions that might help resolve this more quickly:
1. Search for maintainers and people that previously touched the
   related code and @ mention them in a comment.
2. Ask on the [NixOS Discourse](https://discourse.nixos.org/). 3. Ask on the [#nixos channel](irc://irc.freenode.net/#nixos) on
   [irc.freenode.net](https://freenode.net).
```

Co-authored-by: zowoq <59103226+zowoq@users.noreply.github.com>

.github/stale.yml: fix newlines