multipath-tools: Fix build failure due to GZIP make var
The multipath-tools makefiles use GZIP as a variable name but this is
also the name of the environment variable gzip uses to get its default
options.

Normally, this wouldn't get into the environment but nixpkgs exports
GZIP=-n in a setup hook. This in turn causes make to export its own
value for this variable. gzip objects to having -c in the environment
variable and aborts, causing the build to fail.