systemd: Fix cross build
This was very similar to the Mesa issues fixed in
62e6d73a09870174a891d2a8d29a38e62b05efd5: the user-written code is
looking up an unprefixed binutils program.

[I think we should have a way in Meson of specifying a program prefix in
the cross / native files, as a fallback for any program that isn't
explicitly specified. This could both be availible for user written
rules, and help with the default rules.]

Fixes https://github.com/NixOS/mobile-nixos/issues/161