execline: 2.6.0.0 -> 2.6.1.0
 Hello,

 execline-2.6.1.0 is out.

 This version features a more expressive format for the envfile binary.
Most of the files that are suitable for a systemd EnvironmentFile=
directive are now parsable with envfile: double quotes are supported,
backslashed newlines are supported, a few C escape sequences are
supported (including octal and hexadecimal). So it is now possible to
read most existing /etc/default configuration files without spawning a
shell.

 Additionally, envfile now comes with a -I option that makes it ignore
a nonexistent file, instead of failing.

 git://git.skarnet.org/execline
 https://skarnet.org/software/execline/

 Enjoy,
 Bug-reports welcome.