WIP: buildPythonPackage: perform tests in separate derivation
Experiment to see what is needed to perform tests in a separate
derivation.

    $ nix-build -A python3.pkgs.pytest.tests.installation

Ideally, I would like a Python build to consist of the following three
builds:
1) build wheel
2) install wheel
3) test installation
where 3) is optional.

Going there we should probably change the interface significantly,
specifically, we should nomenclature for inputs common to Python users.
All the check related attributes are probably best put in an attribute
set as well, but again, that is a big change in interface.