rmlint: fix running rmlint --gui
rmlint has a `--gui` option which is writing a python "bootstrap" script
in `/tmp`.

It currently can't find its own python modules, as we don't prefix
`PYTHONPATH` to point to it. Also, there's some other runtime
dependencies, mostly due to using gtk from python(3), which also needs
to be preprended to `PYTHONPATH`.

The `rmlint` GUI components also execute `rmlint` again with `--version`
to determine some features, for which we need to prefix `PATH` with
`$out/bin`.