swiften: unbreak package using Arch patch and scons flags
The `swiften` XMPP library was broken as of 2019. This commit
fixes it, by copying a patch from the Arch Linux PKGBUILD
(https://git.archlinux.org/svntogit/community.git/tree/trunk?h=packages/swift)
and by using the same scons flags in that PKGBUILD.
Additionally, we use the SWIFTEN_INSTALLDIR env variable to actually
put something in the output.

Of note is the flag swiften_dll=1, which means that the library
is now built dynamically.