nixosTests.ejabberd: disabling mod_http_upload test
We're adding a flag to xmpp-sendmessage to disable the mod_http_upload
check to let ejabberd use this script with its current configuration.

The current Ejabberd test config does not set up a http_upload
endpoint. Configuring such a endpoint would probably require us to
create a proper TLS setup for the test.

This is not an ideal solution, but at least it'll let us run some
basic test on Ejabberd. Ideally, we should properly setup a
http_upload endpoint in the Ejabberd test to keep it on par with the
prosody one.

We're also pointing passthru.tests to the NixOS VM test to let the
Nixpkgs CI run the test on each Ejabberd update.