Remove Google Talk Plugin
This plugin is no longer necessary anyway, but having it enabled can
cause Firefox and KDE to malfunction, e.g. by hanging for a few
seconds frequently. This is caused by the broken LD_PRELOAD library
that doesn't handle O_TMPFILE properly, so ~/.cache/ksycoca5_* is
created with 0000 permissions. As a result Firefox will constantly
regenerate the ksycoca database.