kitty: 0.17.4 -> 0.18.0
https://github.com/kovidgoyal/kitty/releases/tag/v0.18.0

I submitted a patch to kitty, so that the library paths can now be passed as arguments to the build system, eliminating the need for patching. This should reduce the required maintenance effort in the future.

This commit removes the dependency on `which`, which is no longer necessary since 5e4fe584fab49bfd5d278e513c146b36e76900b5. It was introduced in 481e61fad8b806a48aa30d736344cae4be88dfa6.

The configure phase is now ignored because kitty has no configure script.

The kitty tests are now run after compilation.

The zsh completions no longer need to be invoked with `source` since 38eb73784527587f3fc6df724629b87da66188d0, so remove the outdated comment.

There is now a link to the changelog.