nixos/hardware: adjust readme to the new location
I tried to use `"${pkgs.path}/nixos/hardware/…"` first, but that cause
the module system to go into an infinite recursion.

Ideally we’d have something akin to `modulesPath` (`hardwarePath`?) or
something more general so that we don’t have to depend on
`NIX_PATH` (which is wrong in some circumstances).