Rewrite raspberrypi-builder.sh based on extlinux
This adds the ability to set a configuration limit for the raspberry pi
firmware as well, so we can change the
boot.loader.raspberryPi.uboot.configurationLimit option to
boot.loader.raspberryPi.configurationLimit and use it for both uboot and
the raspberry pi firmware.

It also fixes a bug where not everything in /boot/old would be cleaned
up.

In addition, this passes shellcheck with no errors, and runs shellcheck
as a part of the build.

Co-authored-by: Samuel Dionne-Riel <samuel@dionne-riel.com>