pythonPackages.pyscard: Fix build on Darwin
The previously provided patch is still necessary,
as nix python reports an old version of macOS
that has the bug, when in fact modern macOS
does not have the misspelling.

The patch has been upstreamed, so we take it
to fix 1.9.9 in anticipation of the next release.

(cherry picked from commit 44fd570d7344fb31d7dd92a42d6e1ef872b5f76b)