pythonPackagesOverlays: override all Python package sets at once
Python package sets can be overridden by overriding an interpreter
and passing in `packageOverrides = self: super: {...};`. This is fine
in case you need a single interpreter, however, it does not help you when
you want to override all sets.

With this change it is possible to override all sets at once by appending
a list a list of "overlays" to `pythonPackagesOverlays`.

From reading the implementation you might wonder why a list is used, and not
`lib.composeExtensions`? The reason is the latter requires knowledge of
the library function. This approach should be easier for most users
as it is similar to how we append to lists of e.g. inputs or patches
when overriding a derivation.