chewing-editor: init at 0.1.1
chewing-editor is the user phrase editor
for Chewing input method (a widely used bopomofo Chinese input method)
Chewing IM memorize phrases input by the automatically
or when the hot keys are pressed.
With this package, users will be able to correct/add phrases in the
database. This is one of the fundamental functionality of Chewing IM
and it will be great to have it in NixOS.
The corresponded Packaging Request is #89554.