nixos/openldap: add bootstrapContents option
Adds a bootstrapContents option, which holds LDIF that is used to
initialise the LDAP database if it did not exist previously. Motivation
primarily came from that there is currently a problem where you cannot
start openldap without an existing database, because of slaptest, but
cannot easily create an empty database without running slapd manually.