doc: adding labels doesn't require commit rights
As of [RFC 39][1], package maintainers are members of the NixOS github
organization who don't have commit rights.  However, they *do* have what
GitHub calls [Triage rights][2], which means that package maintainers
are able to add labels to PRs.  The documentation should make this
clear.

I don't know the right way to phrase this in the docs, but the current
text of "Requires commit rights" is definitely incorrect.

[1]: https://discourse.nixos.org/t/join-the-package-maintainer-team/3751
[2]: https://docs.github.com/en/github/setting-up-and-managing-organizations-and-teams/repository-permission-levels-for-an-organization#repository-access-for-each-permission-level